LP bug#677407 / MySQL Bug#48883: Stale data from INNODB_LOCKS table.
The innodb.innodb_information_schema test could fail sporadically due to
the flawed logic in the INFORMATION_SCHEMA.INNODB_LOCKS caching
mechanism.

The logic for how to check when to update the table cache for
INNODB_LOCKS with real data was flawed. This could result in both not
updating the cache often enough (when the table is queried repeatedly
with less than 100 milliseconds in-between) resulting in stale data; as
well as updating too often (when multiple queries against the table
start at around the same time).

Fixed by updating the "last updated" timestamp in the right place, when
the cache is updated, not when it is read.