Bug #692211: innodb_auto_lru_dump crashes if ib_lru_dump doesn't exist
Starting the server with a non-zero innodb_auto_lru_dump value could
crash the server if the dump file does not exist.

The problem was that the 'records' pointer was not initialized at the
start of buf_LRU_file_restore(). This could lead to calling ut_free()
with uninitialized value when the dump file is not found and further
execution of buf_LRU_file_restore() is therefore aborted.

Fixed by initializing 'records' with NULL.