Fixed linker errors about undefined clock_gettime() in client programs when building without SSL support (i.e. without the '--with-ssl' option to configure).
The problem was that in client/Makefile.am "LIBS" is overwritten by
"CLIENT_LIBS". So adding "-lrt" to LIBS alone is not enough, it has to
be added to CLIENT_LIBS as well. --with-ssl pulls the "-lrt" flag to
CLIENT_LIBS, that's why the problem could only be observed in very
specific build configurations (e.g. all BUILD/* scripts did not have
it). Actually, the problem exists in the MySQL server as well. It is
just hidden because MySQL does not check for librt at configure stage,
so it is always built without clock_gettime() support by default.

Fixed by adding "-lrt" to CLIENT_LIBS as well. AC_SEARCH_LIBS() has to
be used instead of AC_CHECK_LIB() now, because apart from adding a
library flag to LIBS we now have to perform an additional action.