Fixes to RPM building scripts
When compiling for an i686 in a CentOS 4 x86_64 machine, rpmbuild doesn't
set correctly _libdir and _arch. Override these.
Allow override of CC, CXX, CFLAGS and CXXFLAGS, for server, HandlerSocket
and UDF.
libmygcc.a is created to the install directories during the %build phase,
so it needs to be stashed at the start of the %install phase, when the
install directories are deleted.