LP bug #764395 / MySQL bug #60788: InnoDB crashes with an assertion                                    failure when receiving a signal on                                    pwrite()
The problem was that InnoDB I/O code was not fully conforming to the
standard on POSIX systems. fsync(), pread() and pwrite() calls may be
interrupted by a signal. In such a case, InnoDB would just fail with an
assertion failure rather than just restarting the interrupted call.

Fixed InnoDB I/O code so that the interrupted system calls are restarted
if they are interrupted by a signal.