Fix LP bug 704216: the option ignore-builtin-innodb is invalid for Percona Server.
Specifying this option for Percona Server either results in a server
running without any InnoDB support, or running with a wrong InnoDB
version.

The fix is to detect this option at the server startup and abort with
an error message.

To accomodate the fix, the test main.innodb_ignore_builtin is made a
no-op.  Ideally we would like to test that the server fails to start
with this option, but that's impossible with the current MTR.