Enable and refresh memory_dynamic_rows.patch with associated tests:
- Regenerate the patch using quilt.

- Fix compilation warnings in hp_extract_record on rec_offsets and
buf_offsets variables.  Verified to be false positive, added
initialization to silence the warning.

- Split out two checks from the percona_heap_bug784464 test into two
new tests: percona_heap_bug784464_32bit and
percona_heap_bug784464_64bit.  This was necessary as the logic for
determining if the variable row format is beneficial for a given
schema considers the fixed and variable chunk sizes in bytes, which
are platform-specific.