Bug #851674: --innodb-optimize-keys does not work correctly with table              without PRIMARY KEY
The problem was that --innodb-optimize-keys tried to "optimize" tables
with implicitly defined primary key, i.e. the ones with no PRIMARY KEY,
but with a UNIQUE key on non-nullable columns. The UNIQUE index picked
as a clustered index by InnoDB was omitted from CREATE TABLE and added
with an ALTER TABLE after dumping the data, which is wrong.

Fixed by taking this case into account by first checking if there is a
primary key defined by querying INFORMATION_SCHEMA. In case there's a
primary key, the first key is always left in CREATE TABLE, whether it's
a PRIMARY KEY or a UNIQUE one. This method relies on the fact that
PRIMARY KEY (if present) is always the first in a SHOW CREATE TABLE
output, and UNIQUE keys always come before others.