Fix for LP bug #860910 / MySQL bug #62557: "SHOW SLAVE STATUS gives wrong output with master-master and using SET uservars"
When a user variable is used in an INSERT/UPDATE query, its value is
written to the binary log as a special User_var event just before the
main query event. The problem was that such User_var events were created
and written with a wrong server_id. The current server_id was always
used even if the main query event came from another server.

As a result, when an INSERT/UPDATE query involving a user variable was
received by a slave, the User_var event was written to slave's binary
log with it's own server_id when log_slave_updates was ON. In a circular
replication setup this led to problems on the server where the original
INSERT/UPDATE query was execute due to Exec_Master_Log_Pos being out of
sync, because it ended up applying the User_var event (which should not
be in the relay log), but not the main query event (which was correctly
skipped). Since User_var events do not have a correct log_pos value,
Exec_Master_Log_Pos was never updated.

Fixed by copying server_id from the original query event when creating
the corresponding User_var events on MYSQL_BIN_LOG::write().