Import Percona Server patch: bug677407.patch
# name       : bug677407
# introduced : 13
# maintainer : Alexey

LP bug#677407 / MySQL Bug#48883: Stale data from INNODB_LOCKS table.

The innodb.innodb_information_schema test could fail sporadically due to
the flawed logic in the INFORMATION_SCHEMA.INNODB_LOCKS caching
mechanism.

The logic for how to check when to update the table cache for
INNODB_LOCKS with real data was flawed. This could result in both not
updating the cache often enough (when the table is queried repeatedly
with less than 100 milliseconds in-between) resulting in stale data; as
well as updating too often (when multiple queries against the table
start at around the same time).

Fixed by updating the "last updated" timestamp in the right place, when
the cache is updated, not when it is read.

I hereby grant Percona the following BSD license to this patch for
inclusion in Percona Server:

Copyright (c) 2010, Kristian Nielsen All rights reserved.

Redistribution and use in source and binary forms, with or without
modification, are permitted provided that the following conditions are
met:

    * Redistributions of source code must retain the above copyright
      notice, this list of conditions and the following disclaimer.
    * Redistributions in binary form must reproduce the above copyright
      notice, this list of conditions and the following disclaimer in
      the documentation and/or other materials provided with the
      distribution.
    * Neither the name of the author nor the names of its contributors
      may be used to endorse or promote products derived from this
      software without specific prior written permission.

THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.