Set correct version in InnoDB log
Down in storage/innodb_plugin/include/univ.i the code that writes the
innodb startup information to the error uses the constant
PERCONA_INNODB_VERSION to include it into the log.

Unless specified in the environment, the constant defaults to 12.5. In
order to make maintenance easy, the build-*.sh scripts were modified so
these set the constant via environment (appending a -D value to CFLAGS).