Fix bug 905711: Crash in DTCollation::set(DTCollation &set), upstream 63793.
The cause and the fix are the same as for upstream 53236 (except that
that testcase does not trigger the crash): don't call member functions
with this == NULL in sql_select.cc:change_cond_ref_to_const.  If
clone_item() returns NULL, then there is no need to set the
collation.

Add new testcase bug63793.