Removal of libmysqlclient-dev packages
libmysqlclient16-dev is a transitional package that depends on
libmysqlclient-dev, which is the package that actually contains all the
development headers and libs.

Since libmysqlclient-dev is also created in the 5.5 branch, which has got
a package libmysqlclient18-dev with this same purpose, the transitional
package has been removed, and libmysqlclient-dev has been renamed to
libmysqlclient16-dev in this branch (5.1). This way, both branches are
able to coexist in the same repository without issues.

In build-dpkg.sh, updated the dch(1) change so it sets the correct
distribution to use in the repository. Previously, unstable was being
used so if the repository didn't contain "unstable" as an alias for the
distribution adding it to the repository would issue an error. This has
been changed to force the actual distribution we're compiling in, so it
works smoothly for both debian and ubuntu.