Update distribution correctly, remove arch-indep files
In build-dpkg.sh, updated the dch(1) change so it sets the correct
distribution to use in the repository. Previously, unstable was being
used so if the repository didn't contain "unstable" as an alias for the
distribution adding it to the repository would issue an error. This has
been changed to force the actual distribution we're compiling in, so it
works smoothly for both debian and ubuntu.