Remove "rel" from the innodb_version variable
Revision 411.2.2 introduced a change in the building scripts so these
began using the PERCONA_INNODB_VERSION they extracted from the Makefile.
However in the Makefile it's reported as relXX.Y but previously in the
code it was used as XX.Y, so this introduced some incompatibilities.

This patch goes back to the XX.Y format.