Pass the includedirs to HandlerSocket via MYSQL_CFLAGS (RPM).
For some installations of rhel5, when mysql-devel 5.0 was installed in the
box at compile time, HandlerSocket would fail to compile.

In ./handlersocket/mysql_incl.hpp, the preprocessor checks the
MYSQL_VERSION_ID (which comes from mysql_version.h) in order to #include
the correct headers. If mysql_version.h is #included from /usr/include
(i.e. from the mysql-devel package) instead of the source being compiled,
an old version would cause the incorrect headers to be included, causing
the compilation to fail.

Prior to this commit, the include dirs of the current source code were
passed via CXXFLAGS, however this caused the intended dirs to be sought
after /usr/include/mysql, which was added by mysql-config --cflags by the
autotools in the HandlerSocket compile phase.

This commit changes it to use MYSQL_CFLAGS, which is prepended by the
configure script to the CXXFLAGS so it's used after the other headers,
causing the correct mysql_version.h to be included.