Revert package name to percona-server-common
The common package had had its name changed to percona-server-common-5.1
as part of other modifications to the control file.

This introduced a change that apt-get upgrade would not solve, since
upgrading would mean it would have to remove the package named
percona-server-common and install the new one. apt-get keeps the upgrades
back in this case and it would only upgrade if dist-upgrade was issued.