Bug #1000221: percona server crash on table import
The problem was that the tablespace had been created with MySQL 5.0
where InnoDB did not set the FIL_PAGE_TYPE field appropriately, so that
field may contain garbage for non-index pages. However, expanded import
relies on that field to detect and update index pages. In this specific
case, it crashed on an extent descriptor page because it assumed it to
be an index page and tried to iterate records on it.

Even though this fix does not resolve the problem completely, it solves
one specific case, i.e. since extent descriptor pages are detected based
on their offset within a tablespace, we know it's not an index page so
there's no reason to process it as an index page.

There's no way to make old data files (i.e. created with MySQL 5.0)
fully supported, there's a separate bug report (#1007264) to document
this.