Fix bug 951588 (Querying I_S.GLOBAL_TEMPORARY_TABLES or TEMPORARY_TABLES crashes threads working with temp tables)
The issue is caused by fill_global_temporary_tables() in sql_show.cc
enumerating existing temp table handler objects that are possibly in
use by other threads.  Then for each object the handler::info() method
is called in store_temporary_table_record() that proceeds to overwrite
current_thd, prebuilt->trx and possibly other fields of any operation
in progress for that temp table handler.

The fix is to not invoke handler::info() method on the existing handler
objects in order not to interfere with their state.  Instead we use
handler::clone() method first to get an independent handler pointing
to the same table and use it to invoke the handler::info() method and
read the stats.

Also simplify the code to get rid of "handle" local variable that
aliases already existing "file".  Fix the type of "engineType" to
const char* and remove unnecessary cast.

For testing add new debug sync point "start_ha_write_row" at the start
of handler::ha_write_row.

Add new test percona_bug_951988 that crashes without the fix and
passes with it.