Fix bug 1008609 (when server is started with both "--userstat=on" and "--bootstrap" paramet
Original title:
Fix bug 1008609 (when server is started with both "--userstat=on" and "--bootstrap" parameters it crashes)

"--userstat=on" and "--bootstrap" parameters it crashes)

Fix bug 1008278 (when slave is started with "--userstat=on"
parameter it crashes)

The issue is caused by derefferencing null pointer thd->net.vio
in sql/sql_connect.cc during checking connection type for
updating ssl connections count in "userstat". If database thread
doesn't have any connection to work with the thd->net.vio
pointer is null.

The fix is to check thd->net.vio pointer for null before checking
connection type in the object reffered by thd->net.vio.

Add new tests percona_bug1008278 and percona_bug1008609 that crash
without the fix and pass with it.