Import mysql-test.diff
-----

mysqldump test would sporadically fail on various hosts (CentOS 5 especially) due to mixing output of stderr and stdout from part of the test. The part of the test only needed to ensure the order of --verbose output, so we instead just capture stderr from mysqldump and discard the data part of the dump.