[enhanced_option_modifiers] In order to offer more control over option visibility, access a
Original title:
[enhanced_option_modifiers] In order to offer more control over option visibility, access and range   limits, the follo

Original title:
In order to offer more control over option visibility, access and range   limits, the following new option modifiers have been added:     minimum-<option name>=<value> : Sets the minimum value that clients       may set <option name> to. If a client does attempt to set the value       of <option name> lesser than <value>, a warning will be issued and       the value of <option name> will be set to <value>.     hidden-<option name>=<TRUE/FALSE> : If set to TRUE, makes any system       variable named in <option name> invisible and immutable to clients.     readonly-<option name>=<TRUE/FALSE> : If set to TRUE, makes any       system variable named in <option name> immutable to clients. Some MySQL option modifiers may be used together in the same option specification, example:   --skip-loose-<name> This set of option modifiers may not be combined together in the same option specification, examples:   OK --loose-readonly-<name>=<T/F>   BAD --readonly-hidden-<name>=<T/F>   BAD --minimum-maximum-<name>=<value>   OK --readonly-<name>=<T/F> --hidden-<name>=<T/F> If an option value has a minimum limit applied and is set to a value lesser than the minimum limit through an option specifier, the value will specified value be kept as the default value. The minimum limit will still apply if a client attempts to set the value lesser than the specified limit. If a client sets an option value to a valid value and the attempts to re-set the value to the out-of-range default, the set limit enforcement will apply and the value will be set to the specified minimum allowed threshold.

  limits, the following new option modifiers have been added:
    minimum-<option name>=<value> : Sets the minimum value that clients
      may set <option name> to. If a client does attempt to set the value
      of <option name> lesser than <value>, a warning will be issued and
      the value of <option name> will be set to <value>.
    hidden-<option name>=<TRUE/FALSE> : If set to TRUE, makes any system
      variable named in <option name> invisible and immutable to clients.
    readonly-<option name>=<TRUE/FALSE> : If set to TRUE, makes any
      system variable named in <option name> immutable to clients.
Some MySQL option modifiers may be used together in the same option
specification, example:
  --skip-loose-<name>
This set of option modifiers may not be combined together in the same
option specification, examples:
  OK --loose-readonly-<name>=<T/F>
  BAD --readonly-hidden-<name>=<T/F>
  BAD --minimum-maximum-<name>=<value>
  OK --readonly-<name>=<T/F> --hidden-<name>=<T/F>
If an option value has a minimum limit applied and is set to a value
lesser than the minimum limit through an option specifier, the value
will specified value be kept as the default value. The minimum limit will
still apply if a client attempts to set the value lesser than the
specified limit. If a client sets an option value to a valid value and
the attempts to re-set the value to the out-of-range default, the set
limit enforcement will apply and the value will be set to the specified
minimum allowed threshold.

==========================================

[enhanced_option_modifiers] Fix for https://bugs.launchpad.net/percona-server/+bug/1167487 - Enhanced option modifiers

Original title:
Fix for https://bugs.launchpad.net/percona-server/+bug/1167487 - Enhanced option modifiers do not deallocate memory | Valgrind: getopt_constraint_get_min_value and getopt_constraint_find definitely lost / still reachable

Originally this was just written off as a known, one time leak, which is is, but the failure causes scripted testing to fail and is really just sloppy.

To fix, a new path needed to be created from mysqld main() through my_cleanup and into the my_getopt.c to properly free the options modifiers structures created and saved during option parsing.
A new function was created called handle_options_end() which is called from cleanup() in mysqld.cc. handle_options_end() calls getopt_constraint_deinit() to signal it that it needs to free its internal static data.
getopt_constraint_init() used to contain a static hash table and initialization flag. These were moved out to global scope so that getopt_constraint_deinit() could have access to them.

While testing, a bad memory read/write was discovered in the min and max constraints which caused a crash during freeing of the constraints structs on shutdown. That has been fixed.

While testing, discovered a bug that caused incorrect behavior, reported as https://bugs.launchpad.net/percona-server/+bug/1233294 - certain option modifiers do not have effect
The my_option_constraints hash is hashed on the option name. Option names are normalized ('-' made '_') before inserting into the hash since 'finds' get called from both main where names have not been transformed and from plugins and other places where the names have been transformed.
The issue is that the searches were performed with non-normalized names, so the find function would not find a record for the same option name (innodb_io_capacity vs innodb-io-capacity). It would then create a new record, transform the name and attempt insertion, which would fail silently due to duplicate.
The new record would be returned and used for that particular constraint setting but would be leaked because it did not exist in the hash. The constraing is then not 'recorded' so any attempt to set outside of the constraint would succeed.
Fixed by transforming the name before looking up the constraint in the hash.