This is implementation if INFORMATION_SCHEMA.INNODB_CHANGED_PAGES table. Blueprint for this feature can be found here:
https://blueprints.launchpad.net/percona-server/+spec/innodb-changed-pages-table
                          
The table is generated from bitmap file which is written by log tracking thread
which parses and writes bitmap blocks after checkpoint happens. Thus the table
contain only information that was written by log tracking thread, it doesn't
contain realtime data.
                          
Information schema tables are in-memory tables, so it can require a lot of
memory to store the whole INFORMATION_SCHEMA.INNODB_CHANGED_PAGES table. There
are two mechanisms to limit the size of the table. The first is condition
pushdown. It means the table will contain only data which corresponds to
condition in "WHERE" SQL clause. And the second is innodb_changed_pages_limit
system variable which limits the number of rows in the table(0 - unlimited).