Fix bug 1012715 (Crash-resistant replication breaks with binlog XA transactin recovery).  T
Original title:
Fix bug 1012715 (Crash-resistant replication breaks with binlog XA transactin recovery).  The cause is that slave position is recorded to InnoDB transaction system header at the XA COMMIT stage.  If a crash happens between XA PREPARE and COMMIT stages, the prepared InnoDB transaction will not have the slave position recorded and thus will fail to update it once it is replayed during binlog crash recovery. The slave log position update is inherent part of changes made by transaction, thus the correct place is the XA PREPARE stage.

transactin recovery).  The cause is that slave position is recorded to
InnoDB transaction system header at the XA COMMIT stage.  If a crash
happens between XA PREPARE and COMMIT stages, the prepared InnoDB
transaction will not have the slave position recorded and thus will
fail to update it once it is replayed during binlog crash recovery.
The slave log position update is inherent part of changes made by
transaction, thus the correct place is the XA PREPARE stage.

Another set of slave positions is still recorded at the XA COMMIT
stage to a different field because the PREPARE-stage write is not
undo-logged and thus is not enough for transaction rollbacks.  The
COMMIT-field value is read on transaction rollback.  This assumes
single-threaded SQL slave (true until MySQL 5.6).

Additional complication is that one master log pos might correspond to
several relay log positions due to relay log rotation.  This is
handled by a kludge where the original positions found in the info
file are saved.  Whenever an overwrite is requested, the new master
log position is compared to the original position, and if they are
equal, then the original relay log position is used for overwrite
instead of stored in InnoDB relay log position.  This works because
the last committed transaction wrote the relay-log.info with the
correct positions, then the file will be overwritten with the last
PREPARED set of positions, then overwritten again with the last
COMMITTED set of positions, which match the original file contents.

- Refactor the relay-log.info overwrite code out of innobase_init() to
  a new function innobase_do_overwrite_relay_log_info() and static
  ha_innodb.cc variables for storing the various log coordinates.  The
  latter removes the five-times-repurposed local vars fname and pos.
  Size the char arrays using symbolic names instead of their values,
  likewise for flag arg of fn_format().  Do file I/O using my_write()
  instead of using IO_CACHE.
- Refactor the code that stores the current MySQL slave log positions
  to an InnoDB trx out of innobase_commit_low() to a new function
  innobase_copy_repl_coords_to_trx().
- Call innobase_copy_repl_coords_to_trx() from innobase_xa_prepare()
  on "real" transaction prepares to save the current slave log
  positions.
- Define new offsets in the trx sys header for the COMMIT set of slave
  log positions.
- New function trx_sys_print_committed_mysql_master_log_pos() that is
  similar to trx_sys_print_mysql_master_log_pos(), but uses the COMMIT
  fields: it reads, saves and prints to stderr the slave log positions
  from the InnoDB trx sys header.  Implement it by factoring out
  trx_sys_read_log_pos() for reading one log coordinate and use it in
  both functions above.
- In innobase_rollback_by_xid() call
  trx_sys_print_commmitted_mysql_master_log_pos() to read the COMMIT
  set of slave log pos and call innobase_do_overwrite_relay_log_info()
  to overwrite the slave info file with them.
- In trx_commit_off_kernel(), write the slave log positions to the
  COMMIT, not PREPARE trx sys fields.
- In trx_prepare_off_kernel(), write the slave log positions to the
  PREPARE trx sys fields.
- Fix typo in the output: relay-info.log instead of the correct
  relay-log.info.
- Store the slave log positions to the current InnoDB transaction in
  innobase_xa_prepare(), not innobase_commit_low().
- Write the slave log positions in the current InnoDB transaction to
  the transaction system header page in trx_prepare_off_kernel()
  instead of trx_commit_off_kernel().
- New crash injection sites "crash_innodb_before_commit" at
  innobase_commit() that triggers just before the "real" transaction
  commits and "crash_innodb_after_prepare" at innobase_xa_prepare()
  right after the "real" transaction prepare.
- New test case rpl_percona_crash_resistant_rpl to test both this bug
  fix and the feature in general.