Fix bug 905334 (Intermittent innodb_bug56680 crash).
The cause is that the buffer pool mutex split added a temporary buffer
block mutex unlock to buf_LRU_free_block() but one its call sites in
buf_page_get_gen() failed to account for the possibility of block
pointing to another page after buf_LRU_free_block() returns due to
this unlock.

Fixed by checking that we indeed have the same page as before the
buf_LRU_free_block() call and retrying the read if that's not the
case.

Adjusted the buf_LRU_free_block() header comment to reflect the
reality.