Fix bug 1038225 (Failing assertion: mutex_own(&buf_pool->mutex) in buf0lru.c on ADD PRIMARY KEY).
The problem is incomplete fix for bug 1015109 which is in turn is a
fix for incorrect merge of upstream revision 0.16394.47.  This
revision introduced new functions, including buf_flush_yield(), that
are called from buf_flush_dirty_pages().  This function acquires the
buffer pool mutex in InnoDB.  In XtraDB it acquires the LRU list mutex
instead.  And buf_flush_yield() in both versions asserts that it
holds, releases and re-reacquires the buffer pool mutex.  This is
incorrect for XtraDB.

Fixed by changing the buf_flush_yield() mutex operations to use the
LRU list mutex.  Adjusted comments for buf_flush_yield() and
buf_flush_try_yield() to reflect XtraDB using the LRU list mutex.