Fix bug 1039384 (innodb_adaptive_flushing_method and innodb_flush_neighbor_pages partially broken).
The symptoms of this bug do not appear in 5.1, but this provides a
starting point of the fix by creating actual tests for the XtraDB enum
plugin variables innodb_adaptive_checkpoint and innodb_read_ahead.
These tests are created in innodb_plugin suite instead of sys_vars one
because the latter is disabled in 5.1 by default.