Fix bug 902471 (Lots of ssl tests failures on oneiric hosts), which is upstream bug http://bugs.mysql.com/bug.php?id=61509.
The crashes are caused by broken inline assembly in taocrypt, that
does not properly pass function args to the inline assembly using the
GCC extended asm syntax, but rather assumes a certain stack layout.
This is not exposed to the compiler and breaks with GCC upgrade and/or
different optimization options.

Workaround by adding -fno-omit-frame-pointer to the taocrypt
compilation flags, thus forcing the required stack layout.  This is
not a proper fix, and might break in the future again.