Have Percona-Server-server depend on Percona-Server-client
Percona-Server-server needs my_check_defaults(1) for the installation, which
was contained in this package. However, in the stock package this program
is included in the client package. Trying to install Percona-Server-server
side by side with the stock client package would result in a conflict for
this program.

The implemented solution is to move my_check_defaults to the
client package, however also we depend from the server directly to the client
package by name. This is to ensure my_print_defaults(1) to be installed,
however we also prevent future incompatibilites with the stock mysql package.

This works with the previous commit which requires the shared package.