Conflict:, Provide: and Replace: mysql-client-core and mysql-server-core
The percona-server-client-5.1 package wasn't either providing nor conflicting
with mysql-client-core-5.1, so trying to install it in an environment where the
latter had been previously installed would result in a problem when running the
install phase, in some cases leaving the dpkg database severely damaged.

The same was happening with the mysql-server-core-5.1 package.

Add a Conflict:, Provide: and Replace: item for mysql-client-core-5.1 and
mysql-server-core-5.1.
Also add a Conflict: line for mysql-client-core-5.5 and mysql-server-core-5.5, 
for the latest releases.