Create a percona-server-dbg-5.1 package for the debugging symbols (dpkg)
Instead of distributing the binaries (such as mysqld) without stripping,
strip them and locate the symbols into a new -dbg package. This is the
equivalent to the -debuginfo packages in rhel -- gdb will find the
symbols for mysqld transparently as long as -dbg is installed. This
packages contains the symbols for all the other packages' binaries.

Increase the debian debhelper compatibility level to 7 in order to take
advante of dh_strip --dbg-package=.