Fix for bug #1049871.
The original patch can be found here:
http://bazaar.launchpad.net/~maria-captains/maria/5.1/revision/3151.1.1 .

The difference between original patch and this one is this patch doesn't
use custom  %`s "printf" specifier for identifiers quoting. Using that specifier
leads to compiller warnings/errors. That warnings/errors could be suppressed
with certain compiler options but such approach can lead to runtime errors
in the code that is not checked on valid printf specifiers. Besides that
the original patch contains two quoting functions (one for mysqlsql string
class and one for mysql prinf functions) which leads to necessity to edit
two places if something should be changed in quoting function.